code-rag/pyproject.toml
[project]
name = "code-rag"
version = "0.1.0"
description = "A lightweight RAG component for retrieving and generating solutions to code errors"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}

dependencies = [
    "chromadb>=0.4.0",            # Vector database
    "datasets>=2.12.0",           # For loading the CodeInsight dataset
    "fastapi>=0.95.0",            # API framework
    "openai>=1.0.0",              # For embeddings
    "pydantic>=2.0.0",            # Data validation
    "python-dotenv>=1.0.0",       # Environment variable management
    "requests>=2.28.0",           # HTTP requests for OpenRouter
    "tenacity>=8.2.0",            # Retry logic for API calls
    "tiktoken>=0.4.0",            # Token counting for OpenAI
    "uvicorn>=0.22.0",            # ASGI server for FastAPI
]

[project.optional-dependencies]
dev = [
    "black>=23.3.0",              # Code formatting
    "isort>=5.12.0",              # Import sorting
    "mypy>=1.3.0",                # Type checking
    "pytest>=7.3.1",              # Testing
    "pytest-cov>=4.1.0",          # Test coverage
]

[build-system]
requires = ["setuptools>=61.0.0", "wheel>=0.40.0"]
build-backend = "setuptools.build_meta"